#1404 Put IIDM nrt diff in a specific file
